requirements/base.txt
Django==1.10.5
django-extensions==1.7.6

django-environ==0.4.1

# MySQL
mysqlclient==1.3.9

# serve files
whitenoise==3.3.0

# Rest Framework
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-filter==1.0.1

# for date processing
python-dateutil==2.6.0

# webpack loading
django-webpack-loader==0.4.1

django-meta==1.3.2
